refactor: move `Agent` from pages folder to components folder
