+++ src/components/agent/agent-type-box/index.tsx
