+++ src/components/agent/availability-box/index.test.tsx
